Fixing POI List Bugs
Firstly this changes adjusts the functionality of the progress bar such
that when you complete one POI you dont nessacarily complete the POI's
before it since they are NOT in order. Additionally I adjusted the
selected poi component to take in the current progress rather then being
hard coded to an order.